refactor(livePage/render): 채팅 렌더 최적화
- ChatWindow.tsx의 내부 로직 의존성 개선
- Chat.tsx에서 채팅 상태를 Input.tsx으로 넘겨 렌더링 비율 감소
- 일부 hover CSS 수정

Resolves: #162
